Extract write.write_datetime to timestamp.rs
No code changes, just a move.
